style: 2 ESLint configs modified
- @stylistic/indent:
This rule is recommended to be disabled by ESLint
author because it aren't very aware about lesser than traditional
Typescript elements. In fact, it pops up in the most random places,
some places hand-formatted, some conflict with formatter(s).

Trivia: it conflicts with both prettier and dprint on indentation.
Considering even dprint doesn't have the neccesary configuration to
amend it, actually pretty(er) amazing.

Alternative: disable the rule inline where it is less than amusing for
the rule to kick in. Disabling a styling rule confuses the reader about
unneccesary diretives, so it isn't a strong option.

- @typescript-eslint/no-unused-vars:
Configure the rule to ignore the _XXX pattern for unused arguments and
variables. This is the TypeScript "noUnusedLocals" configuration
behavior. In fact, _ is considered unnamed/throwaway in a majority of
programming languages.