feat: stable serialization process
The front API for the new serialization (for structuredClone) is in
src/lib/optimizer/new/format.

It relies on user supplied serialization and deserialization method to
safely serialize/deserialize instances to JSON. By default, if a class
is not configured to implement Serializable<SerializedFormat, Class>,
Object.create(clazz.prototype) will be used for deserialization.

Revisited classes that will be serialized to integrate the new API.

test/refactor: Jingliu optimization test suite

- Added 2 new optimizer tests for Jingliu, one with 133 SPD limit.
- Added unit tests for the new serialization API.
- Grouped the test into proper test suites.